Add proper handle of date and time

+++ db-protocol/firebird/src/main/java/org/apache/shardingsphere/db/protocol/firebird/payload/FirebirdPacketPayload.java
@@ -52,7 +52,7 @@ public final class FirebirdPacketPayload implements PacketPayload {
      * @param value 1 byte fixed length integer
      */
     public void writeInt1(final int value) {
-        byteBuf.writeShort(value);
+        byteBuf.writeByte(value);
     }
     
     /**
@@ -70,7 +70,7 @@ public final class FirebirdPacketPayload implements PacketPayload {
      * @param value 2 byte fixed length integer
      */
     public void writeInt2(final int value) {
-        byteBuf.writeInt(value);
+        byteBuf.writeShort(value);
     }
     
     /**
